Split TCanvasElement into its own header
